Rename keywordsInfo.ts to scopedKeywords.ts
